Updated Scripts with new disclaimer
From this point forward, for the scripts that are included within the models I publish, I'll include a short disclaimer at the end of the script to indicate that the model / script was originally published to the Roblox Creator Marketplace by my group (or one of the associated user accounts). I've put this in place so that if my work is ever reuploaded by users with malicious intent, slightly more work would need to be done by them to remove the disclaimer. If a user then happens to use a reuploaded version of my work, it would become slightly easier for them to verify if all the code and etc. was published by a spam account or not. With how problematic reuploading can be in the Creator Marketplace (or at least was in the past), I hope that this'll help combat the few cases where my work could end up being reuploaded in a "virus model" so that users have more of a heads up as to who they're actually getting the model / scripts from.